pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osint-super-kombajn"
version = "0.1.0"
description = "Modułowy system analizy OSINT integrujący klasyczne narzędzia OSINT i sztuczną inteligencję"
readme = "README.md"
authors = [
    {name = "Autor OSINT Super Kombajn", email = "kontakt@przykład.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
]
dependencies = [
    "requests>=2.28.0",
    "aiofiles>=22.1.0",
    "PyYAML>=6.0",
    "jinja2>=3.1.2",
    "httpx>=0.23.3",
    "rich>=13.0.0",
    "tqdm>=4.64.1",
    "phonenumbers>=8.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.1.0",
    "pylint>=2.15.10",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
ai = [
    "transformers>=4.25.0",
    "torch>=2.0.0",
    "langchain>=0.0.154",
]

[project.scripts]
osint-kombajn = "osint_super_kombajn.main:main"

[project.urls]
"Homepage" = "https://github.com/username/OSINT-Super-Kombajn"
"Bug Tracker" = "https://github.com/username/OSINT-Super-Kombajn/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"